requirements.txt
annotated-types==0.5.0
anyio==3.7.0
bcrypt==4.0.1
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.100.1
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orjson==3.9.1
packaging==23.2
Pillow==10.0.1
pluggy==1.3.0
pydantic==2.1.1
pydantic-settings==2.0.3
pydantic_core==2.4.0
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests~=2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.6.3
ujson==5.8.0
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3

